Backend/requirements/requirements.txt
fastapi==0.115.6
pandas==2.2.3
sqlmodel==0.0.22
mysql-connector-python==8.3.0